\name{csi3540RwebApp-package}
\alias{csi3540RwebApp-package}
\alias{csi3540RwebApp}
\docType{package}
\title{
\packageTitle{csi3540RwebApp}
}
\description{
\packageDescription{csi3540RwebApp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{csi3540RwebApp}
\packageIndices{csi3540RwebApp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{csi3540RwebApp}

Maintainer: \packageMaintainer{csi3540RwebApp}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
